+++ limberframework/cache/cache_service_provider.py
@@ -7,14 +7,19 @@ from limberframework.cache.cache import Cache
 from limberframework.cache.lockers import Locker, make_locker
 from limberframework.cache.stores import Store, make_store
 from limberframework.foundation.application import Application
-from limberframework.support.service_providers import ServiceProvider
+from limberframework.support.services import Service, ServiceProvider
 
 
 class CacheServiceProvider(ServiceProvider):
     """Registers cache services to the service container."""
 
-    def register(self):
-        """Registers the cache store to the service container."""
+    def register(self, app: Application):
+        """Registers the cache store to the service container.
+
+        Arguments:
+        app limberframework.foundation.application.Application --
+        the service container.
+        """
 
         async def register_store(app: Application) -> Store:
             """Closure for establishing a cache store.
@@ -25,19 +30,18 @@ class CacheServiceProvider(ServiceProvider):
             Returns Store object.
             """
             config_service = await app.make("config")
-            config = config_service["cache"]
+            config = config_service.get_section("cache")
 
-            # Check whether the cache path is a relative
-            # path, and construct the absolute path.
-            if config_service["cache"]["path"][0] != "/":
-                config = config_service["cache"].copy()
-                config["path"] = (
-                    app.base_path + "/" + config_service["cache"]["path"]
-                )
+            if config["driver"] == "file":
+                config["path"] = app.paths["cache"]
+            elif (
+                config["driver"] == "redis" or config["driver"] == "asyncredis"
+            ) and "password" not in config:
+                config["password"] = None
 
             return await make_store(config)
 
-        self.app.bind("cache.store", register_store, True)
+        app.bind(Service("cache.store", register_store, singleton=True))
 
         async def register_locker(app: Application) -> Locker:
             """Closure for establishing a locker.
@@ -48,13 +52,17 @@ class CacheServiceProvider(ServiceProvider):
             Returns Locker.
             """
             config_service = await app.make("config")
+            config = config_service.get_section("cache")
 
-            if not config_service["cache"]["locker"]:
+            if config["locker"] == "asyncredis" and "password" not in config:
+                config["password"] = None
+
+            try:
+                return await make_locker(config)
+            except ValueError:
                 return None
 
-            return await make_locker(config_service["cache"])
-
-        self.app.bind("cache.locker", register_locker, singleton=True)
+        app.bind(Service("cache.locker", register_locker, singleton=True))
 
         async def register_cache(app: Application) -> Cache:
             """Closure for establishing a
@@ -69,4 +77,4 @@ class CacheServiceProvider(ServiceProvider):
             locker = await app.make("cache.locker")
             return Cache(store, locker)
 
-        self.app.bind("cache", register_cache, defer=True)
+        app.bind(Service("cache", register_cache, defer=True))