+++ limberframework/config/config_service_provider.py
@@ -3,17 +3,24 @@
 Classes:
 - ConfigServiceProvider: Registers configuration services.
 """
+from os import listdir
+from os.path import isfile
+
 from limberframework.config.config import Config
 from limberframework.foundation.application import Application
-from limberframework.support.service_providers import ServiceProvider
+from limberframework.support.services import Service, ServiceProvider
 
 
 class ConfigServiceProvider(ServiceProvider):
     """Registers configuration services to the service container."""
 
-    def register(self) -> None:
+    def register(self, app: Application) -> None:
         """Registers the Config class to the service container,
         which holds configuration settings for the application.
+
+        Arguments:
+        app limberframework.foundation.application.Application --
+        the service container.
         """
 
         async def register_config(app: Application) -> Config:
@@ -25,6 +32,17 @@ class ConfigServiceProvider(ServiceProvider):
             Returns:
             Config -- Config object.
             """
-            return Config()
+            config = Config()
+            config.optionxform = str
 
-        self.app.bind("config", register_config, True)
+            for config_file in listdir(app.paths["config"]):
+                config_file_path = f"{app.paths['config']}/{config_file}"
+
+                if isfile(config_file_path) and config_file.lower().endswith(
+                    ".ini"
+                ):
+                    config.read(config_file_path, encoding="utf-8")
+
+            return config
+
+        app.bind(Service("config", register_config, singleton=True))