+++ limberframework/database/database_service_provider.py
@@ -7,15 +7,19 @@ from sqlalchemy.orm import Session, sessionmaker
 
 from limberframework.database.connections import Connection, make_connection
 from limberframework.foundation.application import Application
-from limberframework.support.service_providers import ServiceProvider
+from limberframework.support.services import Service, ServiceProvider
 
 
 class DatabaseServiceProvider(ServiceProvider):
     """Registers database services to the service container."""
 
-    def register(self) -> None:
+    def register(self, app: Application) -> None:
         """Registers the database connection and session
         services to the service container.
+
+        Arguments:
+        app limberframework.foundation.application.Application --
+        the service container.
         """
 
         async def register_database_connection(app: Application) -> Connection:
@@ -28,9 +32,10 @@ class DatabaseServiceProvider(ServiceProvider):
             Connection object
             """
             config_service = await app.make("config")
-            return await make_connection(config_service["database"])
+            config = config_service.get_section("database")
+            return await make_connection(config)
 
-        self.app.bind("db.connection", register_database_connection)
+        app.bind(Service("db.connection", register_database_connection))
 
         async def register_database_session(app: Application) -> Session:
             """Closure for establishing a database session
@@ -45,4 +50,4 @@ class DatabaseServiceProvider(ServiceProvider):
             db_connection = await app.make("db.connection")
             return sessionmaker(bind=db_connection.engine)()
 
-        self.app.bind("db.session", register_database_session, defer=True)
+        app.bind(Service("db.session", register_database_session, defer=True))