+++ limberframework/foundation/application.py
@@ -4,11 +4,12 @@ Classes:
 - Application: service container that registers and manages services.
 """
 from os import getcwd
+from os.path import join
 from typing import Any
 
 from fastapi import FastAPI
 
-from limberframework.support.service_providers import ServiceProvider
+from limberframework.support.services import Service
 
 
 class Application(FastAPI):
@@ -16,42 +17,43 @@ class Application(FastAPI):
     registering and managing services.
 
     Attributes:
-    config dict -- application configuration settings.
-    bindings dict -- services bound to the service container.
-    instances dict -- instances of singleton services.
+    paths dict - paths to core parts of the application.
+    _bindings dict -- services bound to the service container.
+    _instances dict -- created instances of singleton services.
     """
 
-    def __init__(self, base_path: str = None, *args, **kwargs) -> None:
-        """Establishes the service container."""
-        self.base_path = base_path or getcwd()
-        self.bindings = {}
-        self.instances = {}
+    def __init__(self, *args, base_path: str = None, **kwargs) -> None:
+        """Establishes the service container.
+
+        Arguments:
+        base_path str -- system path to the application.
+        """
+        base_path = base_path or getcwd()
+
+        self.paths = {
+            "base": base_path,
+            "cache": join(base_path, "storage", "cache"),
+            "config": join(base_path, "config"),
+        }
+        self._bindings = {}
+        self._instances = {}
 
         super().__init__(*args, **kwargs)
 
-    def register(self, service_provider: ServiceProvider) -> None:
-        """Register a service provider with the application.
-
-        Arguments:
-        service_provider ServiceProvider -- ServiceProvider object.
-        """
-        service_provider.register()
-
-    def bind(self, name, closure, singleton=False, defer=False) -> None:
+    def bind(self, service: Service) -> None:
         """Bind a service to the application.
 
         Arguments:
-        name str -- name of the service.
-        closure function -- function to call to create the service.
-        singleton bool -- whether multiple instances of the service
-        are allowed.
-        defer bool -- whether to wait loading the service until it is needed.
+        service limberframework.support.services.Service --
+        the service to bind.
         """
-        self.bindings[name] = {
-            "closure": closure,
-            "singleton": singleton,
-            "defer": defer,
-        }
+        if service.name in self._bindings:
+            raise ValueError(
+                f"A service with the name {service.name} has already "
+                f"be bound to the service container."
+            )
+
+        self._bindings[service.name] = service
 
     async def make(self, name: str) -> Any:
         """Create a new instance of a service, if the service
@@ -65,7 +67,7 @@ class Application(FastAPI):
         Any -- an instance of the service.
         """
         try:
-            binding = self.bindings[name]
+            binding = self._bindings[name]
         except KeyError:
             raise KeyError(
                 f"Unknown service {name}, check service "
@@ -73,31 +75,20 @@ class Application(FastAPI):
             )
 
         # If service is not a singleton, return a new instance.
-        if not binding["singleton"]:
-            return await self.bindings[name]["closure"](self)
+        if not binding.singleton:
+            return await self._bindings[name].closure(self)
 
         # If an existing instance of the singleton
         # service is available return it.
-        if name in self.instances.keys():
-            return self.instances[name]
+        if name in self._instances:
+            return self._instances[name]
 
         # Otherwise create a new instance and store it.
-        self.instances[name] = await self.bindings[name]["closure"](self)
-        return self.instances[name]
+        self._instances[name] = await self._bindings[name].closure(self)
+        return self._instances[name]
 
     async def load_services(self) -> None:
         """Make instances of registered services that are not deferrable."""
-        for service, value in self.bindings.items():
-            if not value["defer"]:
-                await self.make(service)
-
-    async def __getitem__(self, name: str) -> Any:
-        """Retrieve a service.
-
-        Arguments:
-        name str -- name of the service.
-
-        Returns:
-        Any -- an instance of the service.
-        """
-        return await self.make(name)
+        for service in self._bindings.values():
+            if not service.defer:
+                await self.make(service.name)