+++ tests/config/test_config.py
@@ -1,24 +1,21 @@
-from pytest import fixture
+from configparser import ConfigParser
+from unittest.mock import Mock
 
 from limberframework.config.config import Config
 
 
-@fixture
-def config():
-    return Config()
+def test_init():
+    config = Config()
+    assert isinstance(config, Config)
+    assert isinstance(config, ConfigParser)
 
 
-def test_init(config):
-    assert config.config == {}
+def test_get_section():
+    config = Config()
+    config.__getitem__ = Mock(
+        return_value={"db": "0", "host": "http://localhost", "port": "1234"}
+    )
 
+    section = config.get_section("test")
 
-def test_get_item(config):
-    config.config["test"] = True
-
-    assert config["test"]
-
-
-def test_set_item(config):
-    config["test"] = True
-
-    assert config.config == {"test": True}
+    assert section == {"db": 0, "host": "http://localhost", "port": 1234}